refactor: 미리보기용 더미 데이터 분리
