+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/compose/team/checklist/TeamChecklistScreen.kt
@@ -30,7 +30,6 @@ import com.bottari.presentation.compose.personal.ChecklistProgressHeader
 import com.bottari.presentation.compose.personal.checklist.PersonalChecklistItem
 import com.bottari.presentation.model.bottari.personal.BottariItemTypeUiModel
 import com.bottari.presentation.model.bottari.team.TeamChecklistItemUiModel
-import kotlin.random.Random
 
 @Composable
 fun TeamChecklistScreen(
@@ -74,7 +73,7 @@ fun TeamChecklistScreen(
     }
 }
 
-fun LazyListScope.teamChecklistNodes(
+private fun LazyListScope.teamChecklistNodes(
     nodes: Map<BottariItemTypeUiModel, Boolean>,
     items: List<TeamChecklistItemUiModel>,
     toggleExpanded: (BottariItemTypeUiModel) -> Unit,
@@ -95,7 +94,7 @@ fun LazyListScope.teamChecklistNodes(
 }
 
 @Composable
-fun TeamChecklistNode(
+private fun TeamChecklistNode(
     node: Map.Entry<BottariItemTypeUiModel, Boolean>,
     items: List<TeamChecklistItemUiModel>,
     toggleExpanded: (BottariItemTypeUiModel) -> Unit,
@@ -203,7 +202,7 @@ private fun SectionHeader(
 @Composable
 private fun TeamChecklistScreenPreview() {
     TeamChecklistScreen(
-        uiState = dummyUiState,
+        uiState = teamChecklistDummyUiState,
         isToolTipClosed = false,
         onCloseToolTip = {},
         onClickSection = {},
@@ -222,34 +221,3 @@ private fun SectionHeaderPreview() {
         modifier = Modifier.fillMaxWidth(),
     )
 }
-
-private fun createDummyProductList(
-    count: Int,
-    type: BottariItemTypeUiModel,
-    idStartIndex: Long = 0,
-): List<TeamChecklistItemUiModel> =
-    List(count) { index ->
-        TeamChecklistItemUiModel(
-            id = idStartIndex + index,
-            name = "더미 아이템 ${idStartIndex + index + 1}",
-            isChecked = Random.nextBoolean(),
-            type = type,
-        )
-    }
-
-private val dummyUiState =
-    ComposeTeamChecklistUiState(
-        isLoading = false,
-        bottariItems =
-            createDummyProductList(
-                10,
-                BottariItemTypeUiModel.SHARED,
-            ) + createDummyProductList(10, BottariItemTypeUiModel.PERSONAL) +
-                createDummyProductList(10, BottariItemTypeUiModel.ASSIGNED()),
-        sections =
-            mapOf(
-                BottariItemTypeUiModel.SHARED to true,
-                BottariItemTypeUiModel.ASSIGNED() to true,
-                BottariItemTypeUiModel.PERSONAL to true,
-            ),
-    )
