+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/compose/team/member/TeamMemberStateCard.kt
@@ -18,21 +18,17 @@ import com.bottari.presentation.R
 import com.bottari.presentation.compose.common.component.BottariBox
 import com.bottari.presentation.compose.common.component.BottariCheckIndicator
 import com.bottari.presentation.compose.common.theme.BottariTheme
-import com.bottari.presentation.model.bottari.PersonalChecklistItemUiModel
 import com.bottari.presentation.model.bottari.team.member.TeamMemberStatusUiModel
-import com.bottari.presentation.model.bottari.team.member.TeamMemberUiModel
 
 @Composable
 fun TeamMemberStateCard(
     memberStatus: TeamMemberStatusUiModel,
-    onClick : () -> Unit,
+    onClick: () -> Unit,
     modifier: Modifier = Modifier,
 ) {
-
     BottariBox(modifier = modifier.clickable(onClick = onClick), contentPadding = PaddingValues(21.dp)) {
         Column {
-            Row (verticalAlignment = Alignment.CenterVertically){
-
+            Row(verticalAlignment = Alignment.CenterVertically) {
                 Text(
                     text = memberStatus.member.nickname,
                     style = BottariTheme.typography.semiBold20.toTextStyle(),
@@ -65,9 +61,10 @@ fun TeamMemberStateCard(
             BottariCheckIndicator(
                 checkedQuantity = memberStatus.checkedItemsCount,
                 totalQuantity = memberStatus.totalItemsCount,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(8.dp),
+                modifier =
+                    Modifier
+                        .fillMaxWidth()
+                        .height(8.dp),
             )
         }
     }
@@ -75,44 +72,9 @@ fun TeamMemberStateCard(
 
 @Preview
 @Composable
-private fun TeamProductStateCardPreview(){
+private fun TeamProductStateCardPreview() {
     TeamMemberStateCard(
-        memberStatus = dummyData[0],
-        onClick = {}
-    )
-}
-
-
-val dummyData = List(10) { index ->
-    val memberId = index + 1
-    val isHost = index == 0
-    val isMe = index == 1
-    val checkedCount = (5..15).random()
-    val totalCount = checkedCount + (2..7).random()
-
-    TeamMemberStatusUiModel(
-        member = TeamMemberUiModel(
-            id = memberId.toLong(),
-            nickname = "멤버 $memberId",
-            isHost = isHost
-        ),
-        totalItemsCount = totalCount,
-        checkedItemsCount = checkedCount,
-        sharedItems = listOf(
-            PersonalChecklistItemUiModel(
-                id = (100 + index).toLong(),
-                name = "공유 아이템 ${index + 1}",
-                isChecked = false
-            )
-        ),
-        assignedItems = listOf(
-            PersonalChecklistItemUiModel(
-                id = (200 + index).toLong(),
-                name = "담당 아이템 ${index + 1}",
-                isChecked = true
-            )
-        ),
-        isMe = isMe,
-        isExpanded = false
+        memberStatus = teamMemberStatusDummyList[0],
+        onClick = {},
     )
 }
